Add files via upload
First update. The project has some error "Field carRepository in fi.eatech.fleetmanagerws.controller.FleetController required a bean of type 'fi.eatech.fleetmanagerws.controller.CarRepository' that could not be found." 
The code and the structure should be practically same as in some tutorials so I'm quite baffled it doesn't work.